@@ -1,12 +1,17 @@
 module github.com/tehcyx/lic
 
+go 1.24
+
 require (
-	github.com/google/go-github/v25 v25.0.2
-	github.com/inconshreveable/mousetrap v1.0.0 // indirect
-	github.com/pelletier/go-toml v1.3.0
-	github.com/spf13/cobra v0.0.3
-	github.com/spf13/pflag v1.0.3 // indirect
-	golang.org/x/oauth2 v0.0.0-20190604053449-0f29369cfe45
+	github.com/google/go-github/v25 v25.1.3
+	github.com/pelletier/go-toml v1.9.5
+	github.com/spf13/cobra v1.8.1
+	golang.org/x/oauth2 v0.24.0
 )
 
-go 1.13
+require (
+	github.com/google/go-cmp v0.6.0 // indirect
+	github.com/google/go-querystring v1.1.0 // indirect
+	github.com/inconshreveable/mousetrap v1.1.0 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
+)
