@@ -1,28 +1,67 @@
 package godep
 
 import (
+	"context"
+	"os"
+	"path/filepath"
 	"testing"
 
 	"github.com/tehcyx/lic/internal/report"
 )
 
 func TestCollect(t *testing.T) {
-	type args struct {
-		proj    *report.Project
-		prjPath string
+	proj := report.NewProjectReport()
+
+	// Create a temporary directory with a test Gopkg.lock file
+	tmpDir := t.TempDir()
+	gopkgPath := filepath.Join(tmpDir, "Gopkg.lock")
+
+	gopkgContent := `# This file is autogenerated, do not edit; changes may be undone by the next 'dep ensure'.
+
+[[projects]]
+  name = "github.com/example/dep1"
+  version = "v1.0.0"
+
+[[projects]]
+  name = "github.com/example/dep2"
+  version = "v2.1.0"
+`
+	if err := os.WriteFile(gopkgPath, []byte(gopkgContent), 0644); err != nil {
+		t.Fatalf("Failed to create test Gopkg.lock: %v", err)
 	}
-	tests := []struct {
-		name    string
-		args    args
-		wantErr bool
-	}{
-		// TODO: Add test cases.
+
+	// Test legacy function
+	err := Collect(proj, tmpDir)
+	if err != nil {
+		t.Errorf("Collect() unexpected error = %v", err)
 	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if err := Collect(tt.args.proj, tt.args.prjPath); (err != nil) != tt.wantErr {
-				t.Errorf("Collect() error = %v, wantErr %v", err, tt.wantErr)
-			}
-		})
+
+	// Test with context cancellation
+	c := NewCollector()
+	ctx, cancel := context.WithCancel(context.Background())
+	cancel() // Cancel immediately
+
+	proj2 := report.NewProjectReport()
+	err = c.Collect(ctx, proj2, tmpDir)
+	if err != context.Canceled {
+		t.Errorf("Collect() with cancelled context should return context.Canceled, got %v", err)
+	}
+
+	// Test CanHandle
+	if !c.CanHandle(tmpDir) {
+		t.Error("CanHandle() should return true for directory with Gopkg.lock")
+	}
+
+	// Test CanHandle with non-existent Gopkg.lock
+	tmpDir2 := t.TempDir()
+	if c.CanHandle(tmpDir2) {
+		t.Error("CanHandle() should return false for directory without Gopkg.lock")
+	}
+}
+
+func TestCollector_Name(t *testing.T) {
+	c := NewCollector()
+	if got := c.Name(); got != "Gopkg.lock" {
+		t.Errorf("Collector.Name() = %v, want 'Gopkg.lock'", got)
 	}
 }
